+++ effects/HealEffect.xaml.cs
@@ -111,8 +111,8 @@ namespace SkillAPITool {
         /// <param name="attribute">attribute</param>
         public void ApplyAttribute(Attribute attribute) {
             if (attribute.Key.EndsWith("Heal")) {
-                damageBaseBox.Text = attribute.Initial.ToString();
-                damageBonusBox.Text = attribute.Scale.ToString();
+                damageBaseBox.Text = ((int)attribute.Initial).ToString();
+                damageBonusBox.Text = ((int)attribute.Scale).ToString();
             }
         }
